rename cmd -> cli
